refactor: ERROR 상수화
