+++ src/validators/CarNameValidator.js
@@ -1,3 +1,4 @@
+import { ERROR_MESSAGE } from "../constants/ErrorMessage.js";
 class CarNameValidator {
   valiateNames(carNames) {
     for (let carName of carNames) {
@@ -11,7 +12,7 @@ class CarNameValidator {
 
   #valiateCarNameLength(carName) {
     if (carName.length < 1 || carName.length > 5) {
-      throw new Error("ERROR : 자동차 이름은 1이상 5이하여야 합니다.");
+      throw new Error(ERROR_MESSAGE.INVALID_NAME_LENGTH);
     }
   }
 
@@ -25,7 +26,7 @@ class CarNameValidator {
 
   #checkDuplicate(carName, nameSet) {
     if (nameSet.has(carName)) {
-      throw new Error("ERROR : 자동차 이름은 중복될 수 없습니다.");
+      throw new Error(ERROR_MESSAGE.DUPLICATED_NAME);
     }
 
     nameSet.add(carName);
@@ -35,15 +36,13 @@ class CarNameValidator {
     const regExp = /^[\{\}\[\]\/?.,;:|\)*~`!^\-_+<>@\#$%&\\\=\(\'\"]$/g;
 
     if (regExp.test(carNames)) {
-      throw new Error(
-        "ERROR : 자동차 이름은 특수기호 만으로 구성될 수 없습니다.",
-      );
+      throw new Error(ERROR_MESSAGE.INVALID_NAME);
     }
   }
 
   #validateCarNamesLength(carNames) {
     if (carNames.length < 2 || carNames.length === 101) {
-      throw new Error("ERROR : 자동차 수는 2이상 100이하여야 합니다.");
+      throw new Error(ERROR_MESSAGE.INVALID_CARS_LENGTH);
     }
   }
 }
