+++ src/validators/TryCountValidator.js
@@ -1,11 +1,12 @@
+import { ERROR_MESSAGE } from "../constants/ErrorMessage.js";
 class TryCountValidator {
   validateNumber(tryCount) {
     if (isNaN(tryCount) === true) {
-      throw new Error("ERROR : 시도 횟수는 숫자여야 합니다.");
+      throw new Error(ERROR_MESSAGE.TRY_COUNT_NOT_NUMBERIC);
     }
 
     if (Number(tryCount) <= 0) {
-      throw new Error("ERROR : 시도 횟수는 양의 정수여야 합니다.");
+      throw new Error(ERROR_MESSAGE.TRY_COUNT_NOT_POSITIVE);
     }
   }
 }
